feat: utils call wrapper
