+++ src/app/context/AuthContext.tsx
@@ -1,5 +1,6 @@
 import { useState, useEffect, createContext, useContext } from "react";
 import utils from "../../common/lib/utils";
+import api from "../../common/lib/api";
 
 interface Account {
   id: string;
@@ -15,24 +16,6 @@ interface AuthContextType {
   getAccountInfo: () => void;
 }
 
-interface AccountInfoRes {
-  currentAccountId: string;
-  info: { alias: string };
-  balance: { balance: string };
-}
-
-interface StatusRes {
-  configured: boolean;
-  unlocked: boolean;
-  currentAccountId: string;
-  info: {
-    alias: string;
-  };
-  balance: {
-    balance: string;
-  };
-}
-
 const AuthContext = createContext<AuthContextType>({} as AuthContextType);
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
@@ -40,12 +23,10 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
   const [loading, setLoading] = useState(true);
 
   const unlock = (password: string, callback: VoidFunction) => {
-    return utils
-      .call<{ currentAccountId: string }>("unlock", { password })
-      .then((response) => {
-        setAccount({ id: response.currentAccountId });
-        callback();
-      });
+    return api.unlock(password).then((response) => {
+      setAccount({ id: response.currentAccountId });
+      callback();
+    });
   };
 
   const lock = (callback: VoidFunction) => {
@@ -59,7 +40,7 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
     const id = account?.id;
     if (!id) return;
     setAccount({ id }); // Clear current info.
-    utils.call<AccountInfoRes>("accountInfo").then((response) => {
+    api.getAccountInfo().then((response) => {
       const { alias } = response.info;
       const balance = parseInt(response.balance.balance); // TODO: handle amounts
       setAccount({ id, alias, balance });
@@ -67,15 +48,15 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
   };
 
   useEffect(() => {
-    utils
-      .call<StatusRes>("status")
+    api
+      .getStatus()
       .then((response) => {
         if (!response.configured) {
           utils.openPage("welcome.html");
           window.close();
         } else if (response.unlocked) {
           setAccount({ id: response.currentAccountId });
-          utils.call<AccountInfoRes>("accountInfo").then((response) => {
+          api.getAccountInfo().then((response) => {
             const { alias } = response.info;
             const balance = parseInt(response.balance.balance); // TODO: handle amounts
             setAccount({ id: response.currentAccountId, alias, balance });
