+++ src/common/lib/api.ts
@@ -1,0 +1,36 @@
+import utils from "./utils";
+
+interface AccountInfoRes {
+  currentAccountId: string;
+  info: { alias: string };
+  balance: { balance: string };
+}
+
+interface StatusRes {
+  configured: boolean;
+  unlocked: boolean;
+  currentAccountId: string;
+  info: {
+    alias: string;
+  };
+  balance: {
+    balance: string;
+  };
+}
+
+interface UnlockRes {
+  currentAccountId: string;
+}
+
+export const getAccountInfo = () => utils.call<AccountInfoRes>("accountInfo");
+
+export const getStatus = () => utils.call<StatusRes>("status");
+
+export const unlock = (password: string) =>
+  utils.call<UnlockRes>("unlock", { password });
+
+export default {
+  getAccountInfo,
+  getStatus,
+  unlock,
+};
